+++ explainerdashboard/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.5.5"
+__version__ = "0.5.6"
 
 from .explainers import ClassifierExplainer, RegressionExplainer  # noqa
 from .dashboards import ExplainerDashboard, ExplainerHub, InlineExplainer  # noqa